pyproject.toml
[project]
name = "nccn-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "brotli>=1.1.0",
    "bs4>=0.0.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.11.0",
    "pypdf>=5.8.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.30.0",
]
